data/scripts/requirements.txt
faker==33.1.0
pandas==2.2.3
supabase==2.13.0
python-dotenv==1.0.1
requests==2.32.3
